pyproject.toml
[project]
name = "task-force-ai"
version = "0.1.0"
description = "Personal multi-agent system accessible via Telegram"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    # Core
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "aiogram>=3.13",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    # Orchestration
    "langgraph>=0.4",
    "langgraph-checkpoint-postgres>=2.0",
    "langchain-core>=0.3",
    # LLM
    "claude-agent-sdk>=0.1.56",
    "anthropic>=0.52",
    # Storage
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",
    "redis>=5.0",
    "alembic>=1.14",
    # Utilities
    "structlog>=24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/taskforce", "config"]
